%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% This file was automatically generated by Rdoc on 2001-04-20 20:17:34.
% Do NOT modify this file, instead modify the source (sma/R/Rinit.R).
% Rdoc is copyright (C) 2001, Henrik Bengtsson, henrikb@braju.com.
% Download the Rdoc compiler at http://www.braju.com/R/
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%


\name{init.addinfo}

\alias{init.addinfo}

\title{Adding Information to a Data Structure for Multi-slide
Microarray Experiments} 

\description{
  Interactive function which adds other information generated from the
  output of image analysis software for microarrays to the existing 
  data structure created using  \code{\link{init.data}}.  
}

\usage{
init.addinfo(batch, attri, dataname=NULL, ...)
}

\arguments{
  \item{batch}{Character string, this refers to the name of a
  collection of experiments.} 
  \item{attri}{Character string, the name of the information to be
    included in the data structure.  For example, from the output of
    \tt{Spot}, this argument can be "area", "signal to noise" etc.
    In other words, these are the column headings from the raw data set.}
   \item{dataname}{A name of your experimental data.  By default it's
      named "batch.exp" where batch is the name of the collection of
      experiments you are interested in.}
   \item{...}{Other arguments}
  }

\value{
 List containing the following component:
  
  \item{R}{contains the raw red intensities, R.}
  \item{G}{contains the raw green intensities, G.}
  \item{Rb}{contains the background red intensities, Rb.}
  \item{Gb}{contains the background green intensities, Gb.}
  as well as information from other users selected columns information.
}

\author{
  Yee Hwa Yang, \email{yeehwa@stat.berkeley.edu} \cr
  Natalie Roberts \email{nroberts@wehi.edu.au}
}

\examples{
## mouse.data <- init.addinfo("mouse", "area")
}

\section{Warning}{The code in the example is not directly executable as
  it draws upon a particular set of data. This data may be downloaded from
\url{http://www.stat.berkeley.edu/users/terry/zarray/Software/smacode.html}
and when loaded appropriately into the user's directory, this example  
should be executable in its current form. }   

%\keyword{microarray, quality information.}

\keyword{manip}

%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% End of File
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
