%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% This file was automatically generated by Rdoc on 2001-04-20 20:17:34.
% Do NOT modify this file, instead modify the source (sma/R/Rinit.R).
% Rdoc is copyright (C) 2001, Henrik Bengtsson, henrikb@braju.com.
% Download the Rdoc compiler at http://www.braju.com/R/
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{init.ctl.index}
\alias{init.ctl.index}
\title{Generates co-ordinates of spots.}
\description{
Generates the 4 co-ordinates of any spots.
}
\usage{
init.ctl.index(grows, gcols, srows, scols)
}
\arguments{
  \item{grows}{The row index of the grid.}
  \item{gcols}{The column index of the grid.}
  \item{srows}{The row index of the spot within the its grid.}
  \item{scols}{The column index of the spot within the its grid.}
}

\value{
  a matrix in which each row contains a vector of 4 integer elements
  which make up the image coordinates of a gene.
}

\examples{
   x <- init.ctl.index(1:4, 1:4, 1:2, 15:20)
## Generates the 4 co-ordinates index to spots in the first 2 rows,
## columns 15 to 20 of every print-tips groups.
}
\keyword{manip}

\author{Yee Hwa Yang, \email{yeehwa@stat.berkeley.edu}}

\note{Sorry: No help files yet}
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% End of File
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
