%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% This file was automatically generated by Rdoc on 2001-04-20 20:17:34.
% Do NOT modify this file, instead modify the source (sma/R/Rinit.R).
% Rdoc is copyright (C) 2001, Henrik Bengtsson, henrikb@braju.com.
% Download the Rdoc compiler at http://www.braju.com/R/
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{init.grid}

\alias{init.grid}

\title{
Initialization of Grid Parameters}

\description{
Interactive function for specifying the dimensions of the spot
matrix and the grid matrix. These parameters depend on the printing
layout of the array, and are used for the within print-tip group
normalization implemented in \code{\link{stat.ma}} and the spatial
representation of spot statistics in \code{\link{plot.spatial}}. 
}

\usage{
init.grid()
}

\arguments{
None.
}

\value{list containing the following components
\item{nspot.r}{ the number of rows of spots per grid;}
\item{nspot.c}{ the number of columns of spots per grid;}
\item{ngrid.r}{ the number of rows of grids per image;}
\item{ngrid.c}{ the number of columns of grids per image.}
}

\references{
Spot manual.
}

\seealso{
\code{\link{plot.mva}}, \code{\link{plot.spatial}},
\code{\link{stat.ma}}, \code{\link{list}}. 
}


\author{
  Yee Hwa Yang, \email{yeehwa@stat.berkeley.edu} \cr
  Sandrine Dudoit, \email{sandrine@stat.berkeley.edu}
}

\examples{
data(MouseArray)
# mouse.setup <- init.grid()

## Here is what you should see:
# Enter number of rows of grids per image (ngrid.r): 4
# Enter number of columns of grids per image (ngrid.c): 4
# Enter number of rows of spots per grid (nspot.r): 19  
# Enter number of columns of spots per grid (nspot.c): 21
# Initialization complete
}

\keyword{misc}
%\keyword{microarray, grid.}
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% End of File
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
