%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% This file was automatically generated by Rdoc on 2001-04-20 20:17:34.
% Do NOT modify this file, instead modify the source (sma/R/Rinit.R).
% Rdoc is copyright (C) 2001, Henrik Bengtsson, henrikb@braju.com.
% Download the Rdoc compiler at http://www.braju.com/R/
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%


\name{init.names}

\alias{init.name.exp}
\alias{init.show.exp}

\title{Set and Read the Names of Experimental Data.}
\description{
  `init.name.exp' creates a look-up table which contains the names of the
  experimental data files and the corresponding object names in R. \cr
  `init.show.exp' displays the look-up table created by
  \code{init.name.exp}.
}

\usage{
init.name.exp(Robject=FALSE)
init.show.exp(batch)
}

\arguments{
 \item{Robject}{if TRUE, the function generates a matrix of characters.
   Otherwise, this matrix is written to a file.}
 \item{batch}{Character string, this refers to the name of a
 collection of experiments.}  
} 

\value{
  \code{init.show.exp} returns a list containing the following components:
  \item{Name in R}{the object names in R;}
  \item{Filename}{the experimental data filenames, including the full
    path name for each file.} 
}
\references{Spot manual
\url{http://www.cmis.csiro.au/iap/Spot/spotmanual.htm}}
  
\author{Yee Hwa Yang, \email{yeehwa@stat.berkeley.edu}}


\examples{
## init.name.exp() ## To create the look-up table.

## This is what you should see:
## Are you creating a new batch.exp file or adding new data names
## to a prexisting batch.exp file? 
## Enter "n" for creating  and "a" for adding new data names: n
## Enter the batch name for the new .exp file: mouse1
## Enter the number of names of files to be entered: 2
##  Enter the R name of your  1 th dataset: m1
##  Enter the actual file name including the full path name for m1 ?
##  ~/path/image1.data 
##  Enter the R name of your  2 th dataset: m2
##  Enter the actual file name including the full path name for m2 ?
##  ~/path/image2.data 
## Finished adding names to .exp file.
## NULL

## View the look-up table.
## init.show.exp("mouse1") 
## 
##    Name in R           Filename
## 1        m1    ~/path/image1.data
## 2        m2    ~/path/image2.data
}

\keyword{attribute}
%\keyword{filename}

%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% End of File
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
