%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% This file was automatically generated by Rdoc on 2001-04-20 20:17:34.
% Do NOT modify this file, instead modify the source (sma/R/Rinit.R).
% Rdoc is copyright (C) 2001, Henrik Bengtsson, henrikb@braju.com.
% Download the Rdoc compiler at http://www.braju.com/R/
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%


\name{init.read.exp}
\alias{init.read.exp}

\title{Reads the Output of the Computed Statistics}

\description{
  Function displays the 30 measurements computed by the program Spot for
  each gene in the slide being considered.} 

\usage{
init.read.exp(batch, imageid, sep="\t", header=TRUE, ...)}

\arguments{
 \item{batch}{batch name of the experiment}
 \item{imageid}{integer value; the index of the slide which is considered}
 \item{sep}{the field separator character; the columns of the file
 will be separated by this character.} 
 \item{header}{a logical value indicating whether the file contains the
   names of the variables as its first line.}
 \item{\dots}{graphical parameters may also be supplied as arguments to
   the function (see \code{\link{par}}).}
}

\value{
  A matrix containing the 30 columns of computed measurements,
  corresponding to the rows of different genes in the specified
  slide. \cr Details regarding these measurements can be found at
  \url{http://www.cmis.csiro.au/iap/Spot/spotoutput.htm}.}

\references{Spot manual \url{http://www.cmis.csiro.au/iap/Spot/spotmanual.htm}}

\author{Yee Hwa Yang, \email{yeehwa@stat.berkeley.edu}}

\examples{
## apoa1.info <- init.read.exp("apoa1", 1) ## obtains the matrix
## of 30 measurements for all the genes spotted on slide 1 of the MouseArray
## experiment.}

\section{Warning}{The code in the example is not directly executable as
  it draws upon a particular set of data. This data may be downloaded from
  \url{http://www.stat.berkeley.edu/users/terry/zarray/Software/smacode.html}
and when loaded appropriately into the user's directory, this
example should be executable in its current form.} 

\keyword{file}
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% End of File
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
