%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% This file was automatically generated by Rdoc on 2001-04-20 20:18:12.
% Do NOT modify this file, instead modify the source (sma/R/Rmisc.R).
% Rdoc is copyright (C) 2001, Henrik Bengtsson, henrikb@braju.com.
% Download the Rdoc compiler at http://www.braju.com/R/
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{is.odd}
\alias{is.odd}
\alias{is.even}

\title{ Determining if a Value is Odd or Even }
\description{
A logical flag which determines if a value supplied by the user is
odd or even. 
}
\usage{
is.odd(x)
is.even(x)
}

\arguments{
 \item{x}{integer value}
}

}
\value{Logical vectors \code{TRUE} or \code{FALSE} are returned
  depending on whether the value is odd or even.

}

\author{ Jessica Mar }

\examples{
is.odd(4)
## FALSE
is.even(100)
## TRUE
}

\keyword{logic}
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% End of File
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
