\name{MouseArray}
\alias{MouseArray}
\alias{mouse1}
\alias{mouse2}
\alias{mouse3}
\alias{mouse4}
\alias{mouse5}
\alias{mouse6}
\alias{mouse.gnames}

\title{Mouse Gene Expression Data from cDNA microarray}

\description{
"MouseArray" consists of 6 different outputs from a cDNA microarray
slide.  The first 3 data (mouse1, mouse2 and mouse3) represent control
hybridization and the other 3 data (mouse4, mouse5 and mouse6)
represent treatment hybridizations.  We have also included vector character
strings representing gene names. 
}

\usage{data(MouseArray)}

\note{For the time being, we are providing a cut-down version of the
  full data set with only partial sets of genes included.  These are
  mainly included to demonstrate the use of the functions in this library.
}

\source{Callow, M. J., Dudoit, S., Gong, E. L., Speed, T. P and Rubin,
  E. M.,  Microarray expression profiling identifies genes with altered 
expression in HDL deficient mice, submitted}

\keyword{datasets}


