%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% This file was automatically generated by Rdoc on 2001-04-20 20:17:34.
% Do NOT modify this file, instead modify the source (sma/R/Rinit.R).
% Rdoc is copyright (C) 2001, Henrik Bengtsson, henrikb@braju.com.
% Download the Rdoc compiler at http://www.braju.com/R/
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
%% Actually RDoc not used here.


\name{read.genepix}

\alias{read.genepix}

\title{Reading in Data Generated by the Image Analysis Package GenePix.}

\description{
Reads in a data file in table format and creates a data frame
with the same number of rows as there are lines in the
 file, and the same number of columns as there are fields
 in the file.\cr
 `read.genepix' reads in the data file generated by the software
 "GenePix".
}

\usage{
read.genepix(name, dir = ".", sep = "\t", header = TRUE, skip = 26, gpname = c("Block", "Col", "Row", "Name", "ID", "X","Y", "Dia", "Rmed", "Rmean", "RSD", "Rbmed", "Rbmean", "RbSD", "Rb1SD", "Rb2SD", "Rbsat","Gmed", "Gmean", "GSD", "Gbmed", "Gbmean", "GbSD", "Gb1SD", "Gb2SD", "Gbsat", "ratiomed", "ratiomean", "medratio", "meanratio", "ratiosd", "Rratio", "RegR2", "FPixels", "BPixels", "summed", "summean", "logratio", "Rmedc", "Gmedc", "Rmeanc", "Gmedc", "flags"),...) 
}

\arguments{
 \item{name}{character string naming the data file from which to read the
data. }
 \item{dir}{character string naming the directory that contains the
 data file.} 

 \item{sep}{the field separator (single  character),  often  "\t"  for
   tab delimited fields.  If omitted,  any  amount  of white space
   (blanks or tabs) can separate the fields.  To read  fixed  format
   files, make  sep  a  numeric vector giving the initial columns of
   the fields. } 

 \item{header}{logical flag: if TRUE, then the first line of the
   file is used as the variable names of the resulting data frame. }

 \item{skip}{the number of lines of the data file to skip before beginning
   to read data.}

 \item{gpname}{A vector of column names. SOme of the column names from
   genepix contain non standard characters, so useful to redefine to
   something R likes to use. Also newer versions of genepix have more columns}

 \item{\dots}{parameters for read.table may also be supplied as arguments to
   the function (see  \code{\link{read.table}}).  }
}

\value{as in \code{\link{read.table}}, a data frame
(\code{\link{data.frame}}) containing a representation of the data
in the file.  
}

\seealso{\code{\link{read.table}}, \code{\link{data.frame}},
\code{\link{read.spot}}.} 

%\keyword{microarray, GenePix.}

\keyword{file}


%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% End of File
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
