%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% This file was automatically generated by Rdoc on 2001-04-20 20:17:34.
% Do NOT modify this file, instead modify the source (sma/R/Rinit.R).
% Rdoc is copyright (C) 2001, Henrik Bengtsson, henrikb@braju.com.
% Download the Rdoc compiler at http://www.braju.com/R/
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{read.spot}

\alias{read.spot}

\title{Reading in Data Generated by the Image Analysis Package Spot}

\description{
Reads in a data file in table format and creates a data frame with the
same number of rows as there are lines in the file, and the same number
of columns as there are fields in the file.\cr
`read.spot': reads in the data file generated by the microarray image
 extraction library Spot. 
}

\usage{
read.spot(name, dir=".", sep="\t", header=TRUE, ...)
}

\arguments{
 \item{name}{character string naming the data file from which to read the
data. }

 \item{dir}{character string naming the directory that contains the data
   file. The default setting is the current directory.}

 \item{sep}{the field separator (single character), often "\t"  for
 tab delimited fields.  If omitted,  any amount of white space
 (blanks or tabs) can separate the fields.  To read fixed format
 files, make  sep  a numeric vector giving the initial columns of
 the fields. } 

 \item{header}{logical flag: if TRUE, then the first line of the
 file is used as the variable names of the resulting data frame. }

 \item{\dots}{parameters for read.table may also be supplied as arguments to
 the function (see  \code{\link{read.table}}).  }
}

\value{as in \code{\link{read.table}}, a data frame
(\code{\link{data.frame}}) containing a representation of the data
in the file.  
}

\seealso{\code{\link{read.table}}, \code{\link{data.frame}},
  \code{\link{write.spot}}, \code{\link{read.genepix}}.}

\examples{

## write.spot(mouse.array, 1, "mouse")
## mouse1 <- read.spot("mouse.1.spot")
}

\keyword{file}
%\keyword{microarray, Spot, GenePix.}

%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% End of File
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
