\name{internal functions}
 
\alias{mouse.lratio}
\alias{mouse.data}
\alias{mouse.setup}
\alias{cl}
\alias{mouse.t2}
\alias{c.min}                
\alias{chen.poly}             
\alias{chur.norm.func}      
\alias{dc.min}                
\alias{in.func}               
\alias{lods.func}             
\alias{nploglikderiv}         
\alias{sq.func}               
\alias{stat.ChurSap.ma}       
\alias{va.func}            
\alias{chen.func.rotate}   
\alias{chen.plot.rotate}  
\alias{chur.M.to.e.func}   
\alias{chur.em.func}    
\alias{chur.pdf}           
\alias{chur.wrapper.func}
\alias{func.em}            
\alias{givelim}    
\alias{lf.func}            
\alias{lod2}       
\alias{newton.plot.rotate}
\alias{nploglik}     
\alias{setup.bayesian}    
\alias{sq2.func}           
\alias{stat.Chen.ma}  
\alias{stat.Newton.ma}     
\alias{va2.func}    
\alias{c.func}
\alias{c.func.est}
\alias{lods.func.est}
\alias{ls.variance}

\title{Internal sma functions}
 
\description{Internal sma functions}
 
\details{These are not to be called by the user.}

\keyword{methods}
