% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sma.analyze}
\alias{sma.analyze}
\title{Analyze splicing microarray data using E*I/J score}
\usage{
sma.analyze(es)
}
\arguments{
\item{es}{ExpressionSet}
}
\value{
list, containing averaged ExpressionSet, score ExpressionSet, statistical analysis based in score ExpressionSet
}
\description{
Averaging of duplicate probes, calculation of score, and statistical analysis
}

