% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sma.load}
\alias{sma.load}
\title{Load raw data files using limma}
\usage{
sma.load(target_path, norm_method = "loess", imputation = "knn")
}
\arguments{
\item{target_path}{character}

\item{norm_method}{character, default is "loess"}

\item{imputation}{character, default is "knn", other values will cause no imputation}
}
\value{
ExpressionSet, normalized Expression set
}
\description{
Loading, normalization, and generation of ExpressionSet using limma functions
}

