% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sma.model}
\alias{sma.model}
\title{Linear modeling of significantly retained introns}
\usage{
sma.model(stats, features, p = 0.05)
}
\arguments{
\item{stats}{list, containing statistical analysis obtained by sma.analyze}

\item{features}{data.frame, containing microarray specific intron features}

\item{p}{numeric, statistical cut-off to be used for analysis, default is FDR = 0.05}
}
\value{
list, containing logistic regression models
}
\description{
Fits logistic regression model to classification into significantly retained and unaffected introns based on intron specific features
}

